board: verdin-am62: allow dfu update
Detect if U-Boot proper was booted from DFU, if so start DFU for
flashing U-boot to eMMC boot area followed by UMS for flashing eMMC
user area.

Upstream-Status: Pending

Initial U-Boot to be used for bring-up and validation of the V1.0
design, we'll decide on the step forward to mainline this once the
bring-up and validation will be done.

Refactoring ideas:

- `Detect DFU boot` moved into arch K3 specific code.
- `Detect DFU boot` by DFUing to a 'DFU' address a magic pattern
  and test that address.
- Have a special U-Boot configuration for DFU boots,.

- With TEZI available for Verdin AM62 this commit can be dropped.
- The TEZI U-Boot could then have in its boot flow:
...
if (! dfu_boot())
	load_the_TEZI_fitimage_from_eMMC();
boot_from_fitimage();
...

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>