board: toradex: verdin-am62: set pmic buck1 +vdd_core to 850 mV
Set PMIC Buck1 aka +VDD_CORE to 850 mV as required to operate the AM625
at its highest operating point of 1.4 GHz.

Upstream-Status: Pending

Initial U-Boot to be used for bring-up and validation of the V1.0
design, we'll decide on the step forward to mainline this once the
bring-up and validation will be done.

Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>