mach-k3: am62_hardware: add lfxosc bypass bit
Add the BP_C bit position of the LFXOSC_CTRL Register.
This allows to bypass the crystal oscillator when providing
the clock from an external source.

Upstream-Status: Pending

Initial U-Boot to be used for bring-up and validation of the V1.0
design, we'll decide on the step forward to mainline this once the
bring-up and validation will be done.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>