verdin-am62: enable the 32kHz clock input
Set the LFOSC0 mode to bypass. This makes the external provided
32kHz clock available.
A noteable consumer is the AM62x internal RTC.

Upstream-Status: Pending

Initial U-Boot to be used for bring-up and validation of the V1.0
design, we'll decide on the step forward to mainline this once the
bring-up and validation will be done.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>