verdin-am62: disable sdhci2
Disable sdhci2 used for the on-module Wi-Fi as distroboot trying to
unsuccessfully initialize it during boot is later causing downstream
Linux to hang during boot.

Upstream-Status: Pending

Initial U-Boot to be used for bring-up and validation of the V1.0
design, we'll decide on the step forward to mainline this once the
bring-up and validation will be done.

Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>