arm: dts: k3-am625-verdin-lpddr4-1600MTs: support up to 95 celsius degrees
Change memory configurations to operate at temperature of 95 degrees.

Currently, dynamic adaptation of temperature related timings is not
supported [1] so the config is for the worst timings.

It is possibile to do so because LPDDR4 devices are refreshed properly
if the memory controller issues REFRESH commands with same or *shorter*
refresh period than reported by MR4 OP[2:0] which depend upon device
temperature.

Configuration is output from SysConfig [2] web tool, currently at version
00.09.08, starting from previous configuration while modifying
these temperature related properties:
 - Operating Temperature Range to "-40C to 95C"
 - tREFIab (ns) to 1950
 - tREFIpb (ns) to 244
 - tRASmax (ns) to 17550

[1] https://e2e.ti.com/support/processors-group/processors/f/processors-forum/1234907/am625-lpddr4-configuration-and-operating-temperature
[2] https://dev.ti.com/sysconfig

Upstream-Status: Pending

Series [3] already send for review with the previous configuration,
we'll send afterward or update it in case a V2 is needed.

[3] https://lore.kernel.org/all/20230607120639.82087-1-marcel@ziswiler.com/

Related-to: ELB-5200
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>