arm64: dts: verdin-am62: add ctrl_sleep_moci# hog
Drive CTRL_SLEEP_MOCI# high at boot (A53 SPL) using a GPIO hog, this
signal may be used to control some power-rails on the carrier board,
therefore it should be set to high when the module is booting.

To do this as early as possible is generally a good idea and the issue
was noticed on the Yavia carrier board where it is needed to power the
I2C EEPROM on the carrier board.

Upstream-Status: Pending

Signed-off-by: Andrejs Cainikovs <andrejs.cainikovs@toradex.com>