arm: dts: k3-*: Add memory node at R5 stage
Add the bootph-pre-ram property to the memory node so that it can be
accessed by FDT functions at R5 stage.

The fdt_setup_mem*() functions require the memory node to be able to
initialize and set the size of the DRAM banks. For this purpose, make
sure all memory nodes are present and standardized, and add them if not.
Also make sure they have bootph-pre-ram property so that it can be
accessible at R5 SPL stage.

Signed-off-by: Neha Malcom Francis <n-francis@ti.com>