configs: am62x: am62px: Enable simple-framebuffer
Enable simple-framebuffer support for AM62x and AM62p platforms which
have early splash screen enabled. This updates the simple-framebuffer
node in Linux device-tree on the fly with framebuffer address and
meta-data and also reserves framebuffer region in device-tree.

Also disable the simple-framebuffer support in am62p's prune
splashscreen config fragment as it is not required without splashscreen
support.

Signed-off-by: Devarsh Thakkar <devarsht@ti.com>